Implements access denied handler and moving package
