+++ api/src/main/kotlin/com/example/todo/application/controller/TodoExceptionHandler.kt
@@ -1,4 +1,4 @@
-package com.example.todo.application.handler
+package com.example.todo.application.controller
 
 import com.example.todo.application.reponse.ResponseData
 import com.example.todo.domain.exception.NotFoundException
